Move step by step assets into docs assets folder
The assets for step by step navigation were previously contained within the `app` folder, which meant that to update it, we'd have to either use our update script to change the contents of the `app` folder, or ask users to do it.

This creates a bit more separation between docs/templates and user apps, moving the assets into the `docs/assets` folder.

Both the app assets and docs assets get compiled to the `/public` directory, so the template views should work in both settings.

This will allow users to continue using their old step by steps seamlessly until they choose to upgrade. Sass doesn't conflict because the new step by step templates use `docs.scss`, which is included after `application.scss`, and I've renamed the JavaScript file `step-by-step-nav.js` to `step-by-step-navigation.js`. This makes it deviate slightly from `govuk_publishing_components`, but I figured it's better to not mess with users' pages.

This is a bit of a stop-gap until we're in a position to make wider architectural changes which separate the docs from the app in a better way.