Explicitly override old styles
Because we use both application.css and docs.css on the template pages, users who still have old CSS files in their app folder will experience some quirks.

This is because there are values in the old styles which are not present in the new styles, so these old values are not overridden.

This PR attempts to manually override each of these instances, in order to work as nicely as possible with this case.

NB: THIS DOESN'T SOLVE THE CASE WHERE THE NEW STYLES OVERRIDE THE OLD STYLES